pyproject.toml
[project]
name = "ai-agents-code-interpreter"
version = "0.1.0"
description = "Hierarchical AI Agents with Python REPL and CSV Analysis tools - From basics (AgentExecutor) to production (LangGraph)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.3",
    "langchain-openai>=1.1.2",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "pandas>=2.3.3",
    "langchainhub>=0.1.21",
    "langchain-experimental>=0.4.1",
    "qrcode>=8.2",
    "pillow>=12.0.0",
    "tabulate>=0.9.0",
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain-tavily>=0.2.14",
    "pydantic>=2.12.5",
    "datetime>=6.0",
]

[dependency-groups]
dev = [
    "ipython>=9.8.0",
    "jupyter>=1.1.1",
    "ruff>=0.14.9",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["agents_basics", "agents_advanced"]
